Fungsi untuk SOAP

+++ src/zzz/progin/tubes4/datastore/Account_has_tugas.java
@@ -8,15 +8,15 @@ import javax.jdo.annotations.Persistent;
 public class Account_has_tugas {
 
 	@Persistent
-	private int account_idAccount;
+	private Long account_idAccount;
 	
 	@Persistent
-	private int tugas_idTugas;
+	private Long tugas_idTugas;
 	
 	@Persistent
 	private boolean isPembuat;
 	
-	public Account_has_tugas(int idAccount, int idTugas, boolean isPembuat){
+	public Account_has_tugas(Long idAccount, Long idTugas, boolean isPembuat){
 		this.setAccount_idAccount(idAccount);
 		this.setTugas_idTugas(idTugas);
 		this.setPembuat(isPembuat);
@@ -30,19 +30,19 @@ public class Account_has_tugas {
 		this.isPembuat = isPembuat;
 	}
 
-	public int getTugas_idTugas() {
+	public Long getTugas_idTugas() {
 		return tugas_idTugas;
 	}
 
-	public void setTugas_idTugas(int tugas_idTugas) {
+	public void setTugas_idTugas(Long tugas_idTugas) {
 		this.tugas_idTugas = tugas_idTugas;
 	}
 
-	public int getAccount_idAccount() {
+	public Long getAccount_idAccount() {
 		return account_idAccount;
 	}
 
-	public void setAccount_idAccount(int account_idAccount) {
+	public void setAccount_idAccount(Long account_idAccount) {
 		this.account_idAccount = account_idAccount;
 	}
 }
