+++ src/zzz/progin/tubes4/datastore/Assignee_has_kategori.java
@@ -8,29 +8,29 @@ import javax.jdo.annotations.Persistent;
 public class Assignee_has_kategori {
 
 	@Persistent
-	private int accounts_idAccount;
+	private Long accounts_idAccount;
 	
 	@Persistent
-	private int kategori_idKategori;
+	private Long kategori_idKategori;
 	
-	public Assignee_has_kategori(int accounts_idAccount, int kategori_idKategori){
+	public Assignee_has_kategori(Long accounts_idAccount, Long kategori_idKategori){
 		this.setAccounts_idAccount(accounts_idAccount);
 		this.setKategori_idKategori(kategori_idKategori);
 	}
 
-	public int getAccounts_idAccount() {
+	public Long getAccounts_idAccount() {
 		return accounts_idAccount;
 	}
 
-	public void setAccounts_idAccount(int accounts_idAccount) {
+	public void setAccounts_idAccount(Long accounts_idAccount) {
 		this.accounts_idAccount = accounts_idAccount;
 	}
 
-	public int getKategori_idKategori() {
+	public Long getKategori_idKategori() {
 		return kategori_idKategori;
 	}
 
-	public void setKategori_idKategori(int kategori_idKategori) {
+	public void setKategori_idKategori(Long kategori_idKategori) {
 		this.kategori_idKategori = kategori_idKategori;
 	}
 }
