+++ src/zzz/progin/tubes4/datastore/Tugas_has_tag.java
@@ -8,29 +8,29 @@ import javax.jdo.annotations.Persistent;
 public class Tugas_has_tag {
 
 	@Persistent
-	private int tugas_idTugas;
+	private Long tugas_idTugas;
 	
 	@Persistent
-	private int tag_idTag;
+	private Long tag_idTag;
 	
-	public Tugas_has_tag(int idTugas, int idTag){
+	public Tugas_has_tag(Long idTugas, Long idTag){
 		this.setTag_idTag(idTag);
 		this.setTugas_idTugas(idTugas);
 	}
 
-	public int getTugas_idTugas() {
+	public Long getTugas_idTugas() {
 		return tugas_idTugas;
 	}
 
-	public void setTugas_idTugas(int tugas_idTugas) {
+	public void setTugas_idTugas(Long tugas_idTugas) {
 		this.tugas_idTugas = tugas_idTugas;
 	}
 
-	public int getTag_idTag() {
+	public Long getTag_idTag() {
 		return tag_idTag;
 	}
 
-	public void setTag_idTag(int tag_idTag) {
+	public void setTag_idTag(Long tag_idTag) {
 		this.tag_idTag = tag_idTag;
 	}
 }
